refactor movegoods for sharing with trade screen
